requirements.txt
# Core dependencies
duckdb>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ML
scikit-learn>=1.3.0

# API & Web
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.0.0

# CLI
click>=8.1.0

# Configuration
PyYAML>=6.0.0

# Data processing
rapidfuzz>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
